F/328 add new db button (#340)
* adds addDatabase to app

* renamen selection

* make drawer always visible

* fix linter errors

* reduce drawer size

* move file in componets

* remove event

* reset data affter dafe

* remove imports

* rename css classes
