+++ hyrisecockpit/frontend/src/App.vue
@@ -1,10 +1,9 @@
 <template>
   <v-app :style="{ background: $vuetify.theme.themes['light'].background }">
-    <AppDrawer :open="showNavigationDrawer" />
+    <AppDrawer />
     <v-app-bar app color="primary" dark>
-      <v-app-bar-nav-icon @click="toggleNavigationDrawer()">
-      </v-app-bar-nav-icon>
       <b> Hyrise Cockpit </b>
+      <add-database />
       <v-icon class="plugin-icon" @click="togglePluginEditor()">
         mdi-widgets
       </v-icon>
@@ -26,6 +25,7 @@
 <script lang="ts">
 import { SetupContext, defineComponent, ref, Ref } from "@vue/composition-api";
 import AppDrawer from "./views/AppDrawer.vue";
+import addDatabase from "./components/addDatabase.vue";
 import PluginsOverview from "./components/plugins/PluginsOverview.vue";
 import WorkloadGeneration from "./components/workload/WorkloadGeneration.vue";
 
@@ -39,7 +39,7 @@ interface Data {
 }
 
 export default defineComponent({
-  components: { AppDrawer, PluginsOverview, WorkloadGeneration },
+  components: { AppDrawer, PluginsOverview, WorkloadGeneration, addDatabase },
   setup(props: {}, context: SetupContext): Data {
     const showNavigationDrawer = ref<boolean>(true);
     const showPluginEditor = ref<boolean>(false);
@@ -69,7 +69,6 @@ export default defineComponent({
 </script>
 <style scoped>
 .plugin-icon {
-  margin-left: auto;
   margin-right: 10px;
 }
 </style>
