+++ hyrisecockpit/frontend/src/plugins/vueRouter.ts
@@ -3,7 +3,6 @@ import VueRouter from "vue-router";
 import Instance from "../views/Instance.vue";
 import InstanceComparison from "../views/InstanceComparison.vue";
 import Overview from "../views/Overview.vue";
-import Selection from "../views/Selection.vue";
 import Home from "../views/Home.vue";
 import WorkloadScreen from "../views/WorkloadScreen.vue";
 
@@ -13,7 +12,6 @@ const routes = [
   { name: "home", path: "/", component: Home },
   { name: "database", path: "/database/:id", component: Instance },
   { name: "workload", path: "/workload", component: WorkloadScreen },
-  { name: "databases", path: "/databases", component: Selection },
   {
     name: "comparison",
     path: "/databases/compare",
