+++ hyrisecockpit/frontend/src/views/AppDrawer.vue
@@ -1,36 +1,28 @@
 <template>
-  <v-navigation-drawer v-model="open" app class="elevation-2" width="200">
+  <v-navigation-drawer v-model="open" app class="elevation-2" width="100">
     <v-toolbar color="primary darken-1"> </v-toolbar>
     <v-card class="text-center" :to="{ name: 'overview' }">
-      <v-icon class="mt-2" size="100">
+      <v-icon class="mt-2" size="70">
         mdi-speedometer
       </v-icon>
-      <v-card-title class="justify-center">
+      <v-card-title class="justify-center card-title">
         Overview
       </v-card-title>
     </v-card>
     <v-card class="text-center" :to="{ name: 'comparison' }">
-      <v-icon class="mt-2" size="100">
+      <v-icon class="mt-2" size="70">
         mdi-database-search
       </v-icon>
-      <v-card-title class="justify-center">
+      <v-card-title class="justify-center card-title">
         Comparison
       </v-card-title>
     </v-card>
     <v-card class="text-center" :to="{ name: 'workload' }">
-      <v-icon class="mt-2" size="100">
+      <v-icon class="mt-2" size="70">
         mdi-account-hard-hat
       </v-icon>
-      <v-card-title class="justify-center">
+      <v-card-title class="justify-center card-title">
         Workload
-      </v-card-title>
-    </v-card>
-    <v-card class="text-center" :to="{ name: 'databases' }">
-      <v-icon class="mt-2" size="100">
-        mdi-cog
-      </v-icon>
-      <v-card-title class="justify-center">
-        Plugins
       </v-card-title>
     </v-card>
   </v-navigation-drawer>
@@ -52,3 +44,9 @@ export default defineComponent({
   }
 });
 </script>
+<style scoped>
+.card-title {
+  padding: 16px 0px 16px 0px;
+  font-size: 16px;
+}
+</style>
